refactor(sip): extract call parking into the ParkOrbit state machine
Third extraction. ParkOrbit owns the orbit slots (park / retrieve /
ring-back / sweep) and the park re-INVITE pending-ACK list; the
attended-transfer ACK list stays with the engine. PbxEnv grows the
engine services parking needs (registration lookup, virtual peers,
session alloc/insert/find, Contact and server-BYE builders), and the
tag/strip header helpers move to SipHeaderUtil.hpp so both translation
units share one definition. Snapshot mirroring keeps the existing
per-site semantics: park ops republish all occupied slots, the tick()
full snapshot lists only Parked ones.
