+++ main/CMakeLists.txt
@@ -79,6 +79,7 @@ idf_component_register(SRCS ${MAIN_SRC}
                             "../src/SIP/TransactionLayer.cpp"
                             "../src/SIP/Registrar.cpp"
                             "../src/SIP/RegisterBeeper.cpp"
+                            "../src/SIP/ParkOrbit.cpp"
                             "../src/SIP/Session.cpp"
                             "../src/SIP/SipClient.cpp"
                             "../src/SIP/SipMessage.cpp"
