+++ tests/CMakeLists.txt
@@ -82,6 +82,7 @@ add_executable(sip_parser_tests
     ../src/SIP/TransactionLayer.cpp
     ../src/SIP/Registrar.cpp
     ../src/SIP/RegisterBeeper.cpp
+    ../src/SIP/ParkOrbit.cpp
     ../src/SIP/SipSdpMessage.cpp
     ../src/SIP/SipClient.cpp
     ../src/SIP/Session.cpp
